requirements.txt
git+https://github.com/msuliot/package.helpers.git
python-dotenv
colorama
rich